removed .view package
