+++ app/src/main/java/com/mmdev/roove/ui/cards/MatchDialogFragment.kt
@@ -1,6 +1,6 @@
 /*
  * Created by Andrii Kovalchuk
- * Copyright (C) 2020. roove
+ * Copyright (C) 2022. roove
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,7 +16,7 @@
  * along with this program.  If not, see https://www.gnu.org/licenses
  */
 
-package com.mmdev.roove.ui.cards.view
+package com.mmdev.roove.ui.cards
 
 import android.os.Bundle
 import android.view.LayoutInflater
