+++ app/src/main/java/com/mmdev/roove/ui/chat/FullScreenDialogFragment.kt
@@ -1,6 +1,6 @@
 /*
  * Created by Andrii Kovalchuk
- * Copyright (C) 2021. roove
+ * Copyright (C) 2022. roove
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,7 +16,7 @@
  * along with this program.  If not, see https://www.gnu.org/licenses
  */
 
-package com.mmdev.roove.ui.chat.view
+package com.mmdev.roove.ui.chat
 
 import android.os.Build
 import android.os.Bundle
